IS: Refactor. Move cronjobs, kafka, and db to infrastrucutre package
