+++ src/main/kotlin/no/nav/syfo/App.kt
@@ -8,17 +8,17 @@ import io.ktor.server.netty.*
 import no.nav.syfo.aktivitetskrav.AktivitetskravService
 import no.nav.syfo.aktivitetskrav.AktivitetskravVarselService
 import no.nav.syfo.aktivitetskrav.VarselPdfService
-import no.nav.syfo.infrastructure.database.AktivitetskravRepository
-import no.nav.syfo.infrastructure.database.AktivitetskravVarselRepository
+import no.nav.syfo.infrastructure.database.repository.AktivitetskravRepository
+import no.nav.syfo.infrastructure.database.repository.AktivitetskravVarselRepository
 import no.nav.syfo.application.ApplicationState
 import no.nav.syfo.application.Environment
 import no.nav.syfo.application.api.apiModule
 import no.nav.syfo.application.cache.RedisStore
-import no.nav.syfo.application.cronjob.launchCronjobModule
-import no.nav.syfo.application.database.applicationDatabase
-import no.nav.syfo.application.database.databaseModule
-import no.nav.syfo.application.kafka.kafkaAivenProducerConfig
-import no.nav.syfo.application.kafka.launchKafkaModule
+import no.nav.syfo.infrastructure.cronjob.launchCronjobModule
+import no.nav.syfo.infrastructure.database.applicationDatabase
+import no.nav.syfo.infrastructure.database.databaseModule
+import no.nav.syfo.infrastructure.kafka.kafkaAivenProducerConfig
+import no.nav.syfo.infrastructure.kafka.launchKafkaModule
 import no.nav.syfo.client.azuread.AzureAdClient
 import no.nav.syfo.client.pdfgen.PdfGenClient
 import no.nav.syfo.infrastructure.pdl.PdlClient
