+++ src/main/kotlin/no/nav/syfo/aktivitetskrav/IAktivitetskravRepository.kt
@@ -1,7 +1,7 @@
 package no.nav.syfo.aktivitetskrav
 
-import no.nav.syfo.infrastructure.database.PAktivitetskrav
-import no.nav.syfo.infrastructure.database.PAktivitetskravVurdering
+import no.nav.syfo.infrastructure.database.repository.PAktivitetskrav
+import no.nav.syfo.infrastructure.database.repository.PAktivitetskravVurdering
 import no.nav.syfo.aktivitetskrav.domain.Aktivitetskrav
 import no.nav.syfo.aktivitetskrav.domain.AktivitetskravVurdering
 import no.nav.syfo.domain.PersonIdent
