+++ src/main/kotlin/no/nav/syfo/application/ApplicationEnvironment.kt
@@ -2,8 +2,8 @@ package no.nav.syfo.application
 
 import io.ktor.server.application.*
 import no.nav.syfo.application.cache.RedisConfig
-import no.nav.syfo.application.database.DatabaseEnvironment
-import no.nav.syfo.application.kafka.KafkaEnvironment
+import no.nav.syfo.infrastructure.database.DatabaseEnvironment
+import no.nav.syfo.infrastructure.kafka.KafkaEnvironment
 import no.nav.syfo.client.ClientEnvironment
 import no.nav.syfo.client.ClientsEnvironment
 import no.nav.syfo.client.OpenClientEnvironment
