+++ src/main/kotlin/no/nav/syfo/application/api/PodApi.kt
@@ -5,7 +5,7 @@ import io.ktor.http.*
 import io.ktor.server.response.*
 import io.ktor.server.routing.*
 import no.nav.syfo.application.ApplicationState
-import no.nav.syfo.application.database.DatabaseInterface
+import no.nav.syfo.infrastructure.database.DatabaseInterface
 
 const val podLivenessPath = "/internal/is_alive"
 const val podReadinessPath = "/internal/is_ready"
