+++ src/main/kotlin/no/nav/syfo/identhendelse/kafka/KafkaIdenthendelseTask.kt
@@ -1,7 +1,8 @@
 package no.nav.syfo.identhendelse.kafka
 
 import no.nav.syfo.application.ApplicationState
-import no.nav.syfo.application.kafka.*
+import no.nav.syfo.infrastructure.kafka.KafkaEnvironment
+import no.nav.syfo.infrastructure.kafka.launchKafkaTask
 
 const val PDL_AKTOR_TOPIC = "pdl.aktor-v2"
