+++ src/main/kotlin/no/nav/syfo/infrastructure/cronjob/CronjobResult.kt
@@ -1,4 +1,4 @@
-package no.nav.syfo.application.cronjob
+package no.nav.syfo.infrastructure.cronjob
 
 data class CronjobResult(
     var updated: Int = 0,
