+++ src/main/kotlin/no/nav/syfo/infrastructure/database/NoElementInsertedException.kt
@@ -1,4 +1,4 @@
-package no.nav.syfo.application.database
+package no.nav.syfo.infrastructure.database
 
 import java.sql.SQLException
