+++ src/main/kotlin/no/nav/syfo/infrastructure/kafka/AktivitetskravVurderingProducer.kt
@@ -3,8 +3,6 @@ package no.nav.syfo.infrastructure.kafka
 import no.nav.syfo.aktivitetskrav.domain.Aktivitetskrav
 import no.nav.syfo.aktivitetskrav.domain.AktivitetskravStatus
 import no.nav.syfo.infrastructure.kafka.domain.KafkaAktivitetskravVurdering
-import no.nav.syfo.application.kafka.KafkaEnvironment
-import no.nav.syfo.application.kafka.kafkaAivenProducerConfig
 import org.apache.kafka.clients.producer.KafkaProducer
 import org.apache.kafka.clients.producer.ProducerConfig
 import org.apache.kafka.clients.producer.ProducerRecord
