+++ src/main/kotlin/no/nav/syfo/infrastructure/kafka/KafkaConsumerService.kt
@@ -1,4 +1,4 @@
-package no.nav.syfo.application.kafka
+package no.nav.syfo.infrastructure.kafka
 
 import org.apache.kafka.clients.consumer.KafkaConsumer
