+++ src/test/kotlin/no/nav/syfo/aktivitetskrav/AktivitetskravServiceSpek.kt
@@ -3,8 +3,8 @@ package no.nav.syfo.aktivitetskrav
 import io.ktor.server.testing.*
 import io.mockk.*
 import kotlinx.coroutines.runBlocking
-import no.nav.syfo.infrastructure.database.AktivitetskravRepository
-import no.nav.syfo.infrastructure.database.AktivitetskravVarselRepository
+import no.nav.syfo.infrastructure.database.repository.AktivitetskravRepository
+import no.nav.syfo.infrastructure.database.repository.AktivitetskravVarselRepository
 import no.nav.syfo.aktivitetskrav.domain.AktivitetskravStatus
 import no.nav.syfo.aktivitetskrav.domain.AktivitetskravVurdering
 import no.nav.syfo.aktivitetskrav.domain.VarselType
